Variable types veranderd in bomb score van uint32_t naar uint16_t en extra comments toegevoegd.

+++ Project_V2/Bomb.h
@@ -24,7 +24,7 @@ public:
 	~Bomb();
 	
 	void drawBomb(uint8_t loc);
-	uint32_t explodeBomb(uint8_t loc);
+	uint16_t explodeBomb(uint8_t loc);
 	void removeAnimation(uint8_t loc);
 	void drawBombAni(uint8_t loc);
 protected:
@@ -34,7 +34,7 @@ private:
 
 	MI0283QT9 *lcdPointer;
 	
-	uint32_t numberOfDestroyed;
+	uint16_t numberOfDestroyed;
 }; //Bomb
 
 #endif //__BOMB_H__
